requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-jinja2==1.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiosqlite==0.20.0
aiounittest==1.4.2
altair==5.4.1
amqp==5.2.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
argon2pure==1.3
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
asyncssh==2.17.0
atpublic==5.0
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
configobj==5.0.9
contourpy==1.3.0
cryptography==43.0.3
cvlib==0.2.7
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docutils==0.21.2
dpath==2.2.0
dulwich==0.22.3
dvc==3.55.2
dvc-data==3.16.6
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
ecdsa==0.19.0
entrypoints==0.4
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
flatbuffers==24.3.25
flatten-dict==0.4.2
flufl.lock==8.1.0
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
funcy==2.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
gmqtt==0.6.11
google-pasta==0.2.0
grandalf==0.8
grpcio==1.67.0
gto==1.7.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
hydra-core==1.3.2
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
imutils==0.5.4
ipykernel==6.29.5
ipython==8.28.0
isoduration==20.11.0
iterative-telemetry==0.0.9
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.6.0
keyring==25.4.1
Kivy==2.3.0
Kivy-Garden==0.1.5
kiwisolver==1.4.7
kombu==5.4.2
libclang==18.1.1
linkify-it-py==2.0.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplot==0.1.9
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
namex==0.0.8
narwhals==1.10.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.1
nh3==0.2.18
notebook==7.2.2
notebook_shim==0.2.4
ntplib==0.4.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.10.0
platformdirs==3.11.0
progressbar==2.5
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydot==3.0.2
pygit2==1.16.0
Pygments==2.18.0
pygtrie==2.5.0
pyloco==0.0.139
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pyroute2==0.7.12
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
readme_renderer==44.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.9.2
rpds-py==0.20.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
scmrepo==3.3.8
SecretStorage==3.3.3
semver==3.0.2
Send2Trash==1.8.3
setuptools==75.2.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
SimpleWebSocketServer==0.1.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
sqltrie==0.11.1
stack-data==0.6.3
streamlit==1.39.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.5.0
terminado==0.18.1
textual==0.84.0
textual-dev==1.6.1
textual-serve==1.1.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.20.1
toml==0.10.2
tomlkit==0.13.2
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.0
triton==3.1.0
twine==5.1.1
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.2.3
ushlex==0.99.1
vine==5.1.0
voluptuous==0.15.2
watchdog==5.0.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.5
wheel==0.44.0
wrapt==1.16.0
xmltodict==0.14.2
yarl==1.15.5
zc.lockfile==3.0.post1
zipp==3.20.2
